src/app/contacts/contact-list/contact-list.component.ts
import { Component, OnInit, OnDestroy } from '@angular/core';
import { Router, ActivatedRoute } from '@angular/router';
import { Subscription } from 'rxjs/Subscription';

import { Contact } from '../../models/contact.model';

import { ContactService } from '../../services/contact.service';

@Component({
  selector: 'app-contact-list',
  templateUrl: './contact-list.component.html',
  styleUrls: ['./contact-list.component.css']
})
export class ContactListComponent implements OnInit, OnDestroy {
    contacts: Contact[];
    subscription: Subscription;

    constructor(private contactService: ContactService,
                private router: Router,
                private route: ActivatedRoute) {
    }

    ngOnInit() {
      this.subscription = this.contactService.contactsChanged
        .subscribe(
          (contacts: Contact[]) => {
            this.contacts = contacts;
          }
        );
      this.contacts = this.contactService.getContacts();
    }

    onNewContact() {
      this.router.navigate(['new'], {relativeTo: this.route});
    }

    ngOnDestroy() {
      this.subscription.unsubscribe();
  }

}
